# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * NBA Prediction Market Trading Bot API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Market {
  id: number;
  polymarketId: string;
  question: string;
  teamA: string;
  /** @nullable */
  teamB: string | null;
  /** series_winner | game_winner | champion | conference */
  marketType: string;
  yesPrice: number;
  noPrice: number;
  volume: number;
  isActive: boolean;
  /** @nullable */
  endDate: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface MarketSummary {
  totalMarkets: number;
  activeMarkets: number;
  totalVolume: number;
  avgEdge: number;
  /** @nullable */
  topOpportunity: string | null;
}

export interface Signal {
  id: number;
  marketId: number;
  marketQuestion: string;
  /** series_arbitrage | momentum | cross_market | nba_stats */
  strategyType: string;
  /** YES | NO */
  direction: string;
  /** Detected edge percentage (0-1) */
  edge: number;
  /** Model confidence (0-1) */
  confidence: number;
  currentPrice: number;
  fairValueEstimate: number;
  /**
   * NBA stat that drove the signal
   * @nullable
   */
  nbaDataPoint: string | null;
  /** pending | executed | expired | rejected */
  status: string;
  createdAt: string;
}

export interface Trade {
  id: number;
  /** @nullable */
  signalId: number | null;
  marketId: number;
  marketQuestion: string;
  direction: string;
  entryPrice: number;
  /** @nullable */
  exitPrice: number | null;
  /** USDC amount */
  size: number;
  kellyFraction: number;
  /** @nullable */
  pnl: number | null;
  /** @nullable */
  pnlPercent: number | null;
  /** open | closed | cancelled */
  status: string;
  openedAt: string;
  /** @nullable */
  closedAt: string | null;
}

export interface EquityPoint {
  timestamp: string;
  bankroll: number;
  pnl: number;
}

export interface PnlSummary {
  totalPnl: number;
  totalPnlPercent: number;
  winRate: number;
  totalTrades: number;
  openTrades: number;
  closedTrades: number;
  bankroll: number;
  startingBankroll: number;
  equityCurve: EquityPoint[];
}

export interface BotStatus {
  isRunning: boolean;
  strategy: string;
  bankroll: number;
  scanIntervalSeconds: number;
  /** @nullable */
  lastScanAt: string | null;
  totalSignalsGenerated: number;
  totalTradesExecuted: number;
  /**
   * Seconds since start
   * @nullable
   */
  uptime: number | null;
}

export interface BotLog {
  id: number;
  /** info | warn | error | signal | trade */
  level: string;
  message: string;
  /** @nullable */
  details: string | null;
  createdAt: string;
}

export type GetMarketsParams = {
  active?: boolean;
};

export type GetSignalsParams = {
  limit?: number;
  strategyType?: string;
};

export type GetTradesParams = {
  limit?: number;
  status?: string;
};

export type GetBotLogsParams = {
  limit?: number;
};