# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: NBA Prediction Market Trading Bot API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /markets:
    get:
      operationId: getMarkets
      tags: [markets]
      summary: Get all tracked NBA prediction markets
      parameters:
        - name: active
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: List of markets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Market"
  /markets/{id}:
    get:
      operationId: getMarket
      tags: [markets]
      summary: Get a single market by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Market details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Market"
        "404":
          description: Not found
...
components:
  schemas:
...
    Market:
      type: object
      properties:
        id:
          type: integer
        polymarketId:
          type: string
        question:
          type: string
        teamA:
          type: string
        teamB:
          type: ["string", "null"]
        marketType:
          type: string
          description: "series_winner | game_winner | champion | conference"
        yesPrice:
          type: number
        noPrice:
          type: number
        volume:
          type: number
        isActive:
          type: boolean
        endDate:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - polymarketId
        - question
        - teamA
        - teamB
        - marketType
        - yesPrice
        - noPrice
        - volume
        - isActive
        - endDate
        - createdAt
        - updatedAt
...