# lib/api-zod/src/generated/types/market.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * NBA Prediction Market Trading Bot API
 * OpenAPI spec version: 0.1.0
 */

export interface Market {
  id: number;
  polymarketId: string;
  question: string;
  teamA: string;
  /** @nullable */
  teamB: string | null;
  /** series_winner | game_winner | champion | conference */
  marketType: string;
  yesPrice: number;
  noPrice: number;
  volume: number;
  isActive: boolean;
  /** @nullable */
  endDate: string | null;
  createdAt: string;
  updatedAt: string;
}